Project Code Changes upto Creating OrderServiceImpl and CreateOrderRequest

+++ src/main/java/com/bikkadit/electronic/store/dtos/CartItemDto.java
@@ -15,7 +15,7 @@ public class CartItemDto {
 
     private Integer quantity;
 
-    private Double totalPrize;
+    private Integer totalPrize;
 
     private CartDto cartDto;
